# Conflicts:
#	app/build.gradle.kts
#	app/src/main/java/org/bxkr/octodiary/DataService.kt
#	app/src/main/java/org/bxkr/octodiary/Utils.kt
#	app/src/main/java/org/bxkr/octodiary/components/Mark.kt
#	app/src/main/java/org/bxkr/octodiary/network/MySchoolLoginService.kt
#	app/src/main/java/org/bxkr/octodiary/screens/NavScreen.kt
#	app/src/main/java/org/bxkr/octodiary/screens/navsections/DaybookScreen.kt
#	app/src/main/java/org/bxkr/octodiary/screens/navsections/MarksScreen.kt

+++ app/src/main/java/org/bxkr/octodiary/models/mark/Activity.kt
@@ -7,5 +7,5 @@ data class Activity(
     @SerializedName("lesson_topic")
     val lessonTopic: String,
     @SerializedName("schedule_item_id")
-    val scheduleItemId: Long
+    val scheduleItemId: Int,
 )
